pyproject.toml
[tool.poetry]
name = "mercury-api"
version = "0.1.0"
description = ""
authors = ["0oAstro <79555780+0oAstro@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.7"
httpx = {extras = ["http2"], version = "^0.28.1"}
pydantic = "^2.10.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
firebase = "^4.0.1"
firebase-admin = "^6.6.0"
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
structlog = "^25.4.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
